Web Inspector: make "sdk" component compile independently from the rest of the front-end. https://bugs.webkit.org/show_bug.cgi?id=94705
Reviewed by Vsevolod Vlasov.

Source/WebCore:

This change refactors the way we compile. Running "compile-front-end.py sdk" will compile sdk only.
It also introduces DOMStorageModel and DatabaseModel that are used to split model / UI layers.

* WebCore.gypi:
* WebCore.vcproj/WebCore.vcproj:
* inspector/InjectedScriptExterns.js:
(InjectedScriptHost.prototype.evaluate):
* inspector/Inspector.json:
* inspector/compile-front-end.py:
(dump_module):
* inspector/front-end/CompilerScriptMapping.js:
* inspector/front-end/ConsoleModel.js:
(WebInspector.ConsoleMessage):
(WebInspector.ConsoleMessage.prototype.clone):
(WebInspector.ConsoleMessage.prototype.location):
* inspector/front-end/DOMAgent.js:
* inspector/front-end/DOMStorage.js:
(WebInspector.DOMStorageModel):
(WebInspector.DOMStorageModel.prototype._addDOMStorage):
(WebInspector.DOMStorageModel.prototype._domStorageUpdated):
(WebInspector.DOMStorageModel.prototype.storageForId):
(WebInspector.DOMStorageModel.prototype.storages):
(WebInspector.DOMStorageDispatcher):
(WebInspector.DOMStorageDispatcher.prototype.addDOMStorage):
(WebInspector.DOMStorageDispatcher.prototype.domStorageUpdated):
* inspector/front-end/Database.js:
(WebInspector.Database):
(WebInspector.Database.prototype.executeSql):
(WebInspector.DatabaseModel):
(WebInspector.DatabaseModel.prototype.databases):
(WebInspector.DatabaseModel.prototype.databaseForId):
(WebInspector.DatabaseModel.prototype._addDatabase):
(WebInspector.DatabaseModel.prototype._sqlTransactionSucceeded):
(WebInspector.DatabaseModel.prototype._sqlTransactionFailed):
(WebInspector.DatabaseDispatcher):
(WebInspector.DatabaseDispatcher.prototype.addDatabase):
(WebInspector.DatabaseDispatcher.prototype.sqlTransactionSucceeded):
(WebInspector.DatabaseDispatcher.prototype.sqlTransactionFailed):
* inspector/front-end/DebuggerModel.js:
* inspector/front-end/FileUtils.js:
(WebInspector.OutputStreamDelegate.prototype.onTransferStarted):
(WebInspector.OutputStreamDelegate.prototype.onChunkTransferred):
(WebInspector.OutputStreamDelegate.prototype.onTransferFinished):
(WebInspector.OutputStreamDelegate.prototype.onError):
(WebInspector.OutputStream):
(WebInspector.OutputStream.prototype.startTransfer):
(WebInspector.OutputStream.prototype.transferChunk):
(WebInspector.OutputStream.prototype.finishTransfer):
(WebInspector.OutputStream.prototype.dispose):
* inspector/front-end/HandlerRegistry.js:
* inspector/front-end/HeapSnapshotView.js:
* inspector/front-end/InspectorFrontendHostStub.js:
* inspector/front-end/PresentationConsoleMessageHelper.js:
(WebInspector.PresentationConsoleMessageHelper.prototype._parsedScriptSource):
* inspector/front-end/ResourcesPanel.js:
(WebInspector.ResourcesPanel):
(WebInspector.ResourcesPanel.prototype._reset):
(WebInspector.ResourcesPanel.prototype._databaseAdded):
(WebInspector.ResourcesPanel.prototype._addDatabase):
(WebInspector.ResourcesPanel.prototype._domStorageAdded):
(WebInspector.ResourcesPanel.prototype._addDOMStorage):
(WebInspector.ResourcesPanel.prototype._showDatabase.get if):
(WebInspector.ResourcesPanel.prototype._showDatabase.else.get if):
(WebInspector.ResourcesPanel.prototype._showDatabase):
(WebInspector.ResourcesPanel.prototype._showDOMStorage.get if):
(WebInspector.ResourcesPanel.prototype._showDOMStorage):
(WebInspector.ResourcesPanel.prototype._updateDatabaseTables.get if):
(WebInspector.ResourcesPanel.prototype._updateDatabaseTables.tableNamesCallback):
(WebInspector.DatabaseTreeElement.prototype.onselect):
(WebInspector.DatabaseTableTreeElement.prototype.onselect):
(WebInspector.DOMStorageTreeElement.prototype.onselect):
* inspector/front-end/ScriptsPanel.js:
(WebInspector.ScriptsPanel.prototype.showFunctionDefinition):
* inspector/front-end/SearchController.js:
(WebInspector.SearchController.prototype.disableSearchUntilExplicitAction):
* inspector/front-end/Settings.js:
* inspector/front-end/Spectrum.js:
* inspector/front-end/TextEditorModel.js:
* inspector/front-end/UIString.js: Added.
(WebInspector.UIString):
* inspector/front-end/UIUtils.js:
* inspector/front-end/WebKit.qrc:
* inspector/front-end/externs.js:
(InspectorFrontendHostAPI.prototype.platform):
(InspectorFrontendHostAPI.prototype.port):
(InspectorFrontendHostAPI.prototype.bringToFront):
(InspectorFrontendHostAPI.prototype.closeWindow):
(InspectorFrontendHostAPI.prototype.requestAttachWindow):
(InspectorFrontendHostAPI.prototype.requestDetachWindow):
(InspectorFrontendHostAPI.prototype.requestSetDockSide):
(InspectorFrontendHostAPI.prototype.setAttachedWindowHeight):
(InspectorFrontendHostAPI.prototype.moveWindowBy):
(InspectorFrontendHostAPI.prototype.setInjectedScriptForOrigin):
(InspectorFrontendHostAPI.prototype.loaded):
(InspectorFrontendHostAPI.prototype.localizedStringsURL):
(InspectorFrontendHostAPI.prototype.hiddenPanels):
(InspectorFrontendHostAPI.prototype.inspectedURLChanged):
(InspectorFrontendHostAPI.prototype.documentCopy):
(InspectorFrontendHostAPI.prototype.copyText):
(InspectorFrontendHostAPI.prototype.openInNewTab):
(InspectorFrontendHostAPI.prototype.canSave):
(InspectorFrontendHostAPI.prototype.save):
(InspectorFrontendHostAPI.prototype.canAppend):
(InspectorFrontendHostAPI.prototype.append):
(InspectorFrontendHostAPI.prototype.sendMessageToBackend):
(InspectorFrontendHostAPI.prototype.recordActionTaken):
(InspectorFrontendHostAPI.prototype.recordPanelShown):
(InspectorFrontendHostAPI.prototype.recordSettingChanged):
(InspectorFrontendHostAPI.prototype.loadResourceSynchronously):
(InspectorFrontendHostAPI.prototype.setZoomFactor):
(InspectorFrontendHostAPI.prototype.canInspectWorkers):
(SourceMapV3):
(SourceMapV3.Section):
(SourceMapV3.Offset):
* inspector/front-end/inspector.html:
* inspector/front-end/inspector.js:
(WebInspector.inspect):

LayoutTests:

* http/tests/inspector-enabled/database-open.html:
* http/tests/inspector-enabled/dom-storage-open.html:

Canonical link: https://commits.webkit.org/112546@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126314 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/inspector/compile-front-end.py
@@ -30,42 +30,32 @@
 import os
 import os.path
 import generate_protocol_externs
+import sys
 
 inspector_path = "Source/WebCore/inspector"
 inspector_frontend_path = inspector_path + "/front-end"
 
 generate_protocol_externs.generate_protocol_externs(inspector_frontend_path + "/protocol-externs.js", inspector_path + "/Inspector.json")
 
-externs = ["externs.js", "protocol-externs.js"]
 jsmodule_name_prefix = "jsmodule_"
 modules = [
     {
-        "target_name": "util",
+        "name": "common",
         "dependencies": [],
         "sources": [
-            "DOMExtension.js",
-            "utilities.js",
-            "treeoutline.js",
-        ]
-    },
-    {
-        "target_name": "common",
-        "dependencies": ["util"],
-        "sources": [
+            "Color.js",
             "Object.js",
             "Settings.js",
+            "UIString.js",
             "UserMetrics.js",
-            "HandlerRegistry.js",
-            "InspectorFrontendHostStub.js",
+            "utilities.js",
         ]
     },
     {
-        "target_name": "sdk",
+        "name": "sdk",
         "dependencies": ["common"],
         "sources": [
-            "InspectorBackend.js",
             "ApplicationCacheModel.js",
-            "Color.js",
             "CompilerScriptMapping.js",
             "ConsoleModel.js",
             "ContentProvider.js",
@@ -85,9 +75,10 @@ modules = [
             "FileUtils.js",
             "HAREntry.js",
             "IndexedDBModel.js",
+            "InspectorBackend.js",
+            "JavaScriptSource.js",
             "Linkifier.js",
             "NetworkLog.js",
-            "Placard.js",
             "PresentationConsoleMessageHelper.js",
             "SASSSourceMapping.js",
             "Script.js",
@@ -109,27 +100,29 @@ modules = [
             "UISourceCode.js",
             "UserAgentSupport.js",
             "Workspace.js",
+            "protocol-externs.js",
         ]
     },
     {
-        "target_name": "ui",
+        "name": "ui",
         "dependencies": ["common"],
         "sources": [
-            "AdvancedSearchController.js",
             "Checkbox.js",
             "ContextMenu.js",
-            "CookiesTable.js",
+            "DOMExtension.js",
             "DOMSyntaxHighlighter.js",
             "DataGrid.js",
             "DefaultTextEditor.js",
             "Dialog.js",
             "Drawer.js",
             "EmptyView.js",
+            "GoToLineDialog.js",
             "HelpScreen.js",
             "InspectorView.js",
             "KeyboardShortcut.js",
             "Panel.js",
             "PanelEnablerView.js",
+            "Placard.js",
             "Popover.js",
             "ProgressBar.js",
             "PropertiesSection.js",
@@ -150,31 +143,33 @@ modules = [
             "TextEditorHighlighter.js",
             "TextEditorModel.js",
             "TextPrompt.js",
+            "TimelineGrid.js",
             "Toolbar.js",
             "UIUtils.js",
             "View.js",
+            "treeoutline.js",
         ]
     },
     {
-        "target_name": "components",
+        "name": "components",
         "dependencies": ["sdk", "ui"],
         "sources": [
+            "AdvancedSearchController.js",
+            "HandlerRegistry.js",
             "ConsoleMessage.js",
+            "CookiesTable.js",
             "DOMBreakpointsSidebarPane.js",
             "DOMPresentationUtils.js",
             "ElementsTreeOutline.js",
-            "GoToLineDialog.js",
             "NativeBreakpointsSidebarPane.js",
             "JavaScriptContextManager.js",
-            "JavaScriptSource.js",
             "ObjectPopoverHelper.js",
             "ObjectPropertiesSection.js",
             "SourceFrame.js",
-            "TimelineGrid.js",
         ]
     },
     {
-        "target_name": "elements",
+        "name": "elements",
         "dependencies": ["components"],
         "sources": [
             "ElementsPanel.js",
@@ -185,7 +180,7 @@ modules = [
         ]
     },
     {
-        "target_name": "network",
+        "name": "network",
         "dependencies": ["components"],
         "sources": [
             "FontView.js",
@@ -205,7 +200,7 @@ modules = [
         ]
     },
     {
-        "target_name": "resources",
+        "name": "resources",
         "dependencies": ["components"],
         "sources": [
             "ApplicationCacheItemsView.js",
@@ -221,7 +216,7 @@ modules = [
         ]
     },
     {
-        "target_name": "scripts",
+        "name": "scripts",
         "dependencies": ["components"],
         "sources": [
             "BreakpointsSidebarPane.js",
@@ -244,7 +239,7 @@ modules = [
         ]
     },
     {
-        "target_name": "console",
+        "name": "console",
         "dependencies": ["components"],
         "sources": [
             "ConsoleView.js",
@@ -252,7 +247,7 @@ modules = [
         ]
     },
     {
-        "target_name": "timeline",
+        "name": "timeline",
         "dependencies": ["components"],
         "sources": [
             "MemoryStatistics.js",
@@ -264,7 +259,7 @@ modules = [
         ]
     },
     {
-        "target_name": "audits",
+        "name": "audits",
         "dependencies": ["components"],
         "sources": [
             "AuditCategories.js",
@@ -276,7 +271,7 @@ modules = [
         ]
     },
     {
-        "target_name": "extensions",
+        "name": "extensions",
         "dependencies": ["components"],
         "sources": [
             "ExtensionAPI.js",
@@ -288,22 +283,28 @@ modules = [
         ]
     },
     {
-        "target_name": "inspector",
+        "name": "settings",
         "dependencies": ["components", "extensions"],
         "sources": [
             "SettingsScreen.js",
+        ]
+    },
+    {
+        "name": "workers",
+        "dependencies": ["components"],
+        "sources": [
             "WorkerManager.js",
         ]
     },
     {
-        "target_name": "tests",
+        "name": "tests",
         "dependencies": ["components"],
         "sources": [
             "TestController.js",
         ]
     },
     {
-        "target_name": "profiler",
+        "name": "profiler",
         "dependencies": ["components"],
         "sources": [
             "BottomUpProfileDataGridTree.js",
@@ -324,33 +325,30 @@ modules = [
             "TopDownProfileDataGridTree.js",
         ]
     },
-#    {
-#        "target_name": "tokenizers",
-#        "dependencies": ["components"],
-#        "sources": [
-#            "SourceCSSTokenizer.js",
-#            "SourceHTMLTokenizer.js",
-#            "SourceJavaScriptTokenizer.js",
-#        ]
-#    },
+    {
+        "name": "host_stub",
+        "dependencies": ["ui"],
+        "sources": [
+            "InspectorFrontendHostStub.js",
+        ]
+    },
 ]
 
-# To be compiled...
-#
-# [Misc]
-# inspector
-# SettingsScreen
-# JavaScriptFormatter
-# ScriptFormatterWorker
-
-compiler_command = "java -jar ~/closure/compiler.jar --summary_detail_level 3 --compilation_level SIMPLE_OPTIMIZATIONS --warning_level VERBOSE --language_in ECMASCRIPT5 --accept_const_keyword \\\n"
-
-command = compiler_command
-for extern in externs:
-    command += "    --externs " + inspector_frontend_path + "/" + extern
-    command += " \\\n"
+modules_by_name = {}
 for module in modules:
-    command += "    --module " + jsmodule_name_prefix + module["target_name"] + ":"
+    modules_by_name[module["name"]] = module
+
+
+def dump_module(name, recursively, processed_modules):
+    if name in processed_modules:
+        return ""
+    processed_modules[name] = True
+    module = modules_by_name[name]
+    command = ""
+    if recursively:
+        for dependency in module["dependencies"]:
+            command += dump_module(dependency, recursively, processed_modules)
+    command += " \\\n    --module " + jsmodule_name_prefix + module["name"] + ":"
     command += str(len(module["sources"]))
     firstDependency = True
     for dependency in module["dependencies"]:
@@ -360,31 +358,49 @@ for module in modules:
             command += ","
         firstDependency = False
         command += jsmodule_name_prefix + dependency
-    command += " \\\n"
     for script in module["sources"]:
-        command += "        --js " + inspector_frontend_path + "/" + script
-        command += " \\\n"
-command += "\n"
-os.system(command)
+        command += " \\\n        --js " + inspector_frontend_path + "/" + script
+    return command
 
-print "Compiling InjectedScriptSource.js..."
-os.system("echo \"var injectedScriptValue = \" > " + inspector_path + "/" + "InjectedScriptSourceTmp.js")
-os.system("cat  " + inspector_path + "/" + "InjectedScriptSource.js" + " >> " + inspector_path + "/" + "InjectedScriptSourceTmp.js")
-command = compiler_command
-command += "    --externs " + inspector_path + "/" + "InjectedScriptExterns.js" + " \\\n"
-command += "    --module " + jsmodule_name_prefix + "injected_script" + ":" + "1" + " \\\n"
-command += "        --js " + inspector_path + "/" + "InjectedScriptSourceTmp.js" + " \\\n"
-command += "\n"
-os.system(command)
-os.system("rm " + inspector_path + "/" + "InjectedScriptSourceTmp.js")
+compiler_command = "java -jar ~/closure/compiler.jar --summary_detail_level 3 --compilation_level SIMPLE_OPTIMIZATIONS --warning_level VERBOSE --language_in ECMASCRIPT5 --accept_const_keyword \\\n"
 
-print "Compiling InjectedScriptWebGLModuleSource.js..."
-os.system("echo \"var injectedScriptWebGLModuleValue = \" > " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js")
-os.system("cat  " + inspector_path + "/" + "InjectedScriptWebGLModuleSource.js" + " >> " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js")
-command = compiler_command
-command += "    --externs " + inspector_path + "/" + "InjectedScriptExterns.js" + " \\\n"
-command += "    --module " + jsmodule_name_prefix + "injected_script" + ":" + "1" + " \\\n"
-command += "        --js " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js" + " \\\n"
-command += "\n"
-os.system(command)
-os.system("rm " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js")
+process_recursively = len(sys.argv) == 2
+if process_recursively:
+    module_name = sys.argv[1]
+    if module_name != "all":
+        modules = [modules_by_name[sys.argv[1]]]
+    for module in modules:
+        command = compiler_command
+        command += "    --externs " + inspector_frontend_path + "/externs.js"
+        command += dump_module(module["name"], True, {})
+        print "Compiling \"" + module["name"] + "\""
+        os.system(command)
+else:
+    command = compiler_command
+    command += "    --externs " + inspector_frontend_path + "/externs.js"
+    for module in modules:
+        command += dump_module(module["name"], False, {})
+    os.system(command)
+
+if not process_recursively:
+    print "Compiling InjectedScriptSource.js..."
+    os.system("echo \"var injectedScriptValue = \" > " + inspector_path + "/" + "InjectedScriptSourceTmp.js")
+    os.system("cat  " + inspector_path + "/" + "InjectedScriptSource.js" + " >> " + inspector_path + "/" + "InjectedScriptSourceTmp.js")
+    command = compiler_command
+    command += "    --externs " + inspector_path + "/" + "InjectedScriptExterns.js" + " \\\n"
+    command += "    --module " + jsmodule_name_prefix + "injected_script" + ":" + "1" + " \\\n"
+    command += "        --js " + inspector_path + "/" + "InjectedScriptSourceTmp.js" + " \\\n"
+    command += "\n"
+    os.system(command)
+    os.system("rm " + inspector_path + "/" + "InjectedScriptSourceTmp.js")
+
+    print "Compiling InjectedScriptWebGLModuleSource.js..."
+    os.system("echo \"var injectedScriptWebGLModuleValue = \" > " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js")
+    os.system("cat  " + inspector_path + "/" + "InjectedScriptWebGLModuleSource.js" + " >> " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js")
+    command = compiler_command
+    command += "    --externs " + inspector_path + "/" + "InjectedScriptExterns.js" + " \\\n"
+    command += "    --module " + jsmodule_name_prefix + "injected_script" + ":" + "1" + " \\\n"
+    command += "        --js " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js" + " \\\n"
+    command += "\n"
+    os.system(command)
+    os.system("rm " + inspector_path + "/" + "InjectedScriptWebGLModuleSourceTmp.js")
